requirements.txt
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7
python-dotenv>=1.0
